feat(chip): Introduce enhanced chip and collection management
- Added a new `builder` module for improved ergonomics in chip creation.
- Implemented `ChipCollection` for managing groups of chips with selection modes.
- Created a `Chip` core module for state management and validation.
- Developed a `view` module for rendering chips with Material Design styling.
- Reorganized chip-related functionality into focused submodules for better maintainability.
- Added convenience functions for quick chip creation.
- Implemented comprehensive tests for chip creation, selection, and validation.
