+++ abop-gui/src/styling/material/components/selection/mod.rs
@@ -83,10 +83,15 @@ pub mod tests;
 
 // Re-export core types and traits (but not convenience functions)
 pub use builder::{
-    BatchBuilder, Checkbox, CheckboxBuilder, Chip, ChipBuilder, ComponentBuilder,
+    BatchBuilder, Checkbox, CheckboxBuilder, ComponentBuilder,
     ConditionalBuilder, Radio, RadioBuilder, Switch, SwitchBuilder,
 };
-pub use chip::{ChipCollection, ChipCollectionBuilder, ChipSelectionMode};
+// Updated chip exports to use the new modular structure
+pub use chip::{
+    Chip, ChipBuilder, ChipCollection, ChipCollectionBuilder, ChipSelectionMode,
+    MAX_CHIP_LABEL_LENGTH, DEFAULT_ANIMATION_DURATION,
+    filter_chip_collection, single_select_chip_collection,
+};
 pub use common::{
     AnimatedWidget, AnimationConfig, CheckboxState, ChipState, ChipVariant, ComponentProps,
     ComponentSize, EasingCurve, SelectionError, SelectionWidget, StatefulWidget, SwitchState,
